Conflicts:
	src/main/resources/applicationContext-mvc.xml
	src/main/resources/shopxx.properties

+++ src/main/java/net/shopxx/controller/admin/AreaController.java
@@ -19,7 +19,9 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import net.shopxx.Message;
 import net.shopxx.entity.Area;
+import net.shopxx.entity.Country;
 import net.shopxx.service.AreaService;
+import net.shopxx.service.CountryService;
 
 /**
  * Controller - 地区
@@ -33,22 +35,50 @@ public class AreaController extends BaseController {
 
 	@Inject
 	private AreaService areaService;
+	
+	@Inject
+    private CountryService countryService;
 
 	/**
 	 * 添加
 	 */
 	@GetMapping("/add")
-	public String add(Long parentId, ModelMap model) {
+	public String add(Long parentId,Long countryId, ModelMap model) {
+	    //添加国家
+	    if (parentId == null && countryId == null) {
+	        return "admin/area/addCountry";
+	    }
+	    if (countryId != null) {
+	      Country c =  countryService.find(countryId);
+	      model.addAttribute("country", c);
+	      return "admin/area/add";
+	    }
 		model.addAttribute("parent", areaService.find(parentId));
 		return "admin/area/add";
 	}
+	
+	
+	/**
+     * 保存国家
+     */
+    @PostMapping("/saveCountry")
+    public String saveCountry(Country country, Long parentId, RedirectAttributes redirectAttributes) {
+        country.setState(1);    
+        countryService.save(country);
+            return "redirect:list";
+    }
 
 	/**
 	 * 保存
 	 */
 	@PostMapping("/save")
-	public String save(Area area, Long parentId, RedirectAttributes redirectAttributes) {
-		area.setParent(areaService.find(parentId));
+	public String save(Area area, Long parentId, Long countryId, RedirectAttributes redirectAttributes) {
+	    if (parentId != null) {
+	        area.setParent(areaService.find(parentId));
+	    }
+	    if (countryId != null) {
+	        area.setCountry(countryService.find(countryId));
+	    }
 		if (!isValid(area)) {
 			return ERROR_VIEW;
 		}
@@ -62,18 +92,27 @@ public class AreaController extends BaseController {
 		area.setDeliveryCenters(null);
 		area.setFreightConfigs(null);
 		areaService.save(area);
-		addFlashMessage(redirectAttributes, SUCCESS_MESSAGE);
+		addFlashMessage(redirectAttributes, Message.success(SUCCESS_MESSAGE));
 		return "redirect:list";
 	}
 
 	/**
 	 * 编辑
 	 */
-	@GetMapping("/edit")
-	public String edit(Long id, ModelMap model) {
-		model.addAttribute("area", areaService.find(id));
-		return "admin/area/edit";
+	@GetMapping("/editCountry")
+	public String editCountry(Long id, ModelMap model) {
+		model.addAttribute("country", countryService.find(id));
+		return "admin/area/editCountry";
 	}
+	
+	 /**
+     * 编辑地区
+     */
+    @GetMapping("/edit")
+    public String edit(Long id, ModelMap model) {
+        model.addAttribute("area", areaService.find(id));
+        return "admin/area/edit";
+    }
 
 	/**
 	 * 更新
@@ -83,22 +122,45 @@ public class AreaController extends BaseController {
 		if (!isValid(area)) {
 			return ERROR_VIEW;
 		}
-		areaService.update(area, "fullName", "treePath", "grade", "parent", "children", "members", "receivers", "orders", "deliveryCenters", "freightConfigs");
-		addFlashMessage(redirectAttributes, SUCCESS_MESSAGE);
+		areaService.update(area, "fullName", "treePath", "grade", "parent", "children", "members", "receivers", "orders", "deliveryCenters", "freightConfigs","country");
+		addFlashMessage(redirectAttributes, Message.success(SUCCESS_MESSAGE));
 		return "redirect:list";
 	}
+	
+
+    /**
+     * 更新
+     */
+    @PostMapping("/updateCountry")
+    public String update(Country country, RedirectAttributes redirectAttributes) {
+        if (!isValid(country)) {
+            return ERROR_VIEW;
+        }
+        countryService.update(country, "state", "version");
+        addFlashMessage(redirectAttributes, Message.success(SUCCESS_MESSAGE));
+        return "redirect:list";
+    }
 
 	/**
 	 * 列表
 	 */
 	@GetMapping("/list")
-	public String list(Long parentId, ModelMap model) {
+	public String list(Long parentId, Long countryId,ModelMap model) {
+	    //国家
+	    if (countryId == null && parentId == null) {
+	        model.addAttribute("countries", countryService.findRoots()); 
+	        return "admin/area/list";
+	    }
+	    model.addAttribute("countries", null); 
+	   
 		Area parent = areaService.find(parentId);
 		if (parent != null) {
 			model.addAttribute("parent", parent);
 			model.addAttribute("areas", new ArrayList<>(parent.getChildren()));
 		} else {
-			model.addAttribute("areas", areaService.findRoots());
+		    Country c =  countryService.find(countryId);
+		    model.addAttribute("country",  c);
+			model.addAttribute("areas", new ArrayList<>(c.getAreas()));
 		}
 		return "admin/area/list";
 	}
@@ -107,9 +169,15 @@ public class AreaController extends BaseController {
 	 * 删除
 	 */
 	@PostMapping("/delete")
-	public @ResponseBody Message delete(Long id) {
-		areaService.delete(id);
-		return SUCCESS_MESSAGE;
+	public @ResponseBody Message delete(Long id,Long country) {
+	    if (country !=null && country == 1) {
+	       Country c =  countryService.find(id);
+	       c.setState(0);
+	       countryService.update(c);
+	    } else {
+	        areaService.delete(id);
+	    }
+		return Message.success(SUCCESS_MESSAGE);
 	}
 
 }
